Extract view helpers into SystemInfo module
